Update CI script.
