+++ .github/workflows/code_test_and_deploy.yml
@@ -57,17 +57,18 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install .[dev]
-      - name: Test
+      - name: Test All
         # run SSH tests only on Linux because Windows and macOS
         # are already run within a virtual container and so cannot
         # run Linux containers because nested containerisation is disabled.
+      - name: Test SSH (Linux only)
+        if: runner.os == 'Linux'
         run: |
-          if [ "$RUNNER_OS" == "Linux" ]; then
-             sudo service mysql stop  # free up port  3306 for ssh tests: https://github.com/orgs/community/discussions/25550
-             pytest -k "test_combinations_ssh_transfer or test_ssh_setup"
-          else
-             pytest -k "not test_combinations_ssh_transfer and not test_ssh_setup"
-          fi
+          sudo service mysql stop  # free up port 3306 for ssh tests
+          pytest tests/tests_transfers/ssh
+      - name: Test All
+        run: |
+          pytest --ignore tests/tests_transfers/ssh
 
   build_sdist_wheels:
     name: Build source distribution
