--- tests/quick_make_project.py
@@ -1,5 +1,0 @@
-base_path = r"C:/Users/Joe/work/git-repos/forks/yxtuix/joe"
-
-from test_utils import quick_create_project
-
-quick_create_project(base_path)
