+++ tests/test_utils.py
@@ -10,7 +10,6 @@ from os.path import join
 from pathlib import Path
 
 import yaml
-from file_conflicts_pathtable import get_pathtable
 
 from datashuttle import DataShuttle
 from datashuttle.configs import canonical_configs, canonical_folders
@@ -151,19 +150,6 @@ def setup_project_fixture(tmp_path, test_project_name, project_type="full"):
 
 def make_test_path(base_path, local_or_central, test_project_name):
     return Path(base_path) / local_or_central / test_project_name
-
-
-def create_all_pathtable_files(pathtable):
-    """ """
-    for i in range(pathtable.shape[0]):
-        filepath = pathtable["base_folder"][i] / pathtable["path"][i]
-        filepath.parents[0].mkdir(parents=True, exist_ok=True)
-        write_file(filepath, contents="test_entry")
-
-
-def quick_create_project(base_path):
-    pathtable = get_pathtable(base_path)
-    create_all_pathtable_files(pathtable)
 
 
 # -----------------------------------------------------------------------------
