+++ tests/tests_integration/test_formatting.py
@@ -1,8 +1,9 @@
 import pytest
-from base import BaseTest
 
 from datashuttle.utils import formatting
 from datashuttle.utils.custom_exceptions import NeuroBlueprintError
+
+from ..base import BaseTest
 
 
 class TestFormatting(BaseTest):
