+++ tests/tests_integration/test_local_only_mode.py
@@ -1,12 +1,13 @@
 import shutil
 
 import pytest
-import test_utils
-from base import BaseTest
 
 from datashuttle.utils.custom_exceptions import (
     ConfigError,
 )
+
+from .. import test_utils
+from ..base import BaseTest
 
 TEST_PROJECT_NAME = "test_project"
