+++ tests/tests_integration/test_validation.py
@@ -3,11 +3,12 @@ import shutil
 import warnings
 
 import pytest
-from base import BaseTest
 
 from datashuttle import quick_validate_project
 from datashuttle.utils import formatting, validation
 from datashuttle.utils.custom_exceptions import NeuroBlueprintError
+
+from ..base import BaseTest
 
 # -----------------------------------------------------------------------------
 # Inconsistent sub or ses value lengths
