+++ tests/tests_regression/test_backwards_compatibility.py
@@ -3,7 +3,10 @@ import shutil
 from pathlib import Path
 
 import pytest
-import test_utils
+
+from datashuttle import DataShuttle
+
+from .. import test_utils
 
 TEST_PROJECT_NAME = "test_project"
