+++ tests/tests_transfers/ssh/ssh_test_images/Dockerfile
