+++ tests/tests_transfers/ssh/test_ssh_setup.py
@@ -3,11 +3,12 @@ import copy
 import platform
 
 import pytest
-import ssh_test_utils
-import test_utils
-from base_transfer import BaseTransfer
 
 from datashuttle.utils import ssh
+
+from ... import test_utils
+from . import ssh_test_utils
+from .base_ssh import BaseSSHTransfer
 
 TEST_SSH = ssh_test_utils.get_test_ssh()
 
@@ -16,7 +17,7 @@ TEST_SSH = ssh_test_utils.get_test_ssh()
     platform.system == "Darwin", reason="Docker set up is not robust on macOS."
 )
 @pytest.mark.skipif(not TEST_SSH, reason="TEST_SSH is false")
-class TestSSH(BaseTransfer):
+class TestSSH(BaseSSHTransfer):
 
     @pytest.fixture(scope="function")
     def project(test, tmp_path, setup_ssh_container):
