+++ tests/tests_transfers/ssh/test_ssh_transfer.py
@@ -3,10 +3,11 @@ import shutil
 
 import paramiko
 import pytest
-import ssh_test_utils
-from base_transfer import BaseTransfer
 
 from datashuttle.utils import ssh
+
+from . import ssh_test_utils
+from .base_ssh import BaseSSHTransfer
 
 TEST_SSH = ssh_test_utils.get_test_ssh()
 
@@ -15,7 +16,7 @@ TEST_SSH = ssh_test_utils.get_test_ssh()
     platform.system == "Darwin", reason="Docker set up is not robust on macOS."
 )
 @pytest.mark.skipif(not TEST_SSH, reason="TEST_SSH is false")
-class TestSSHTransfer(BaseTransfer):
+class TestSSHTransfer(BaseSSHTransfer):
 
     @pytest.fixture(
         scope="class",
