+++ tests/tests_tui/test_tui_logging.py
@@ -1,9 +1,12 @@
 import pytest
+
 import test_utils
 from tui_base import TuiBase
 
 from datashuttle.tui.app import TuiApp
 from datashuttle.tui.tabs.logging import RichLogScreen
+
+from .tui_base import TuiBase
 
 
 class TestTuiLogging(TuiBase):
