+++ tests/tests_tui/test_tui_validate.py
@@ -2,10 +2,11 @@ import warnings
 
 import pytest
 import textual
-from tui_base import TuiBase
 
 import datashuttle
 from datashuttle.tui.app import TuiApp
+
+from .tui_base import TuiBase
 
 
 class TestTuiValidate(TuiBase):
