+++ tests/tests_tui/tui_base.py
@@ -1,10 +1,11 @@
 import pytest_asyncio
-import test_utils
 from textual.widgets._tabbed_content import ContentTab
 
 from datashuttle.configs import canonical_configs
 from datashuttle.tui.screens.project_manager import ProjectManagerScreen
 from datashuttle.tui.screens.project_selector import ProjectSelectorScreen
+
+from .. import test_utils
 
 
 class TuiBase:
